requirements.txt
requests
sys
os
json
time
threading
PyQt5
python-xlib
